requirement.txt
streamlit
langchain
langchain-community
google-generativeai
python-dotenv
pandas
huggingface_hub

faiss-cpu
langchain_google_genai
PyPDF2
faiss-cpu
InstructorEmbedding
sentence-transformers
pypdf




